zoom double to float

+++ NNControl/Network/NeuralNetworkController.cs
@@ -64,7 +64,7 @@ namespace NNControl.Network
         }
 
 
-        public void SetZoom(double value)
+        public void SetZoom(float value)
         {
             if (value < -1.0d)
             {
